requirements.txt
absl-py==0.11.0
apipkg==1.5
argon2-cffi @ file:///C:/ci/argon2-cffi_1613037959010/work
astroid==2.4.2
astunparse==1.6.3
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
atomicwrites==1.4.0
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bleach @ file:///tmp/build/80754af9/bleach_1612211392645/work
cachetools==4.1.1
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1613247279197/work
chardet==3.0.4
click==7.1.2
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
contextlib2==0.6.0.post1
cycler==0.10.0
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
defusedxml @ file:///home/ktietz/src/ci_mi/defusedxml_1612808095953/work
easydict==1.9
entrypoints==0.3
execnet==1.8.0
ffmpeg==1.4
flatbuffers==1.12
gast==0.3.3
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
httplib2==0.18.1
idna==2.10
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
iniconfig==1.1.1
install==1.3.4
ipykernel==5.5.0
ipython @ file:///C:/ci/ipython_1614616640087/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.6.3
isort==5.7.0
jedi==0.17.0
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
joblib==0.17.0
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core @ file:///C:/ci/jupyter_core_1612213356021/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver @ file:///C:/ci/kiwisolver_1612282606037/work
lazy-object-proxy==1.4.3
lxml @ file:///C:/ci/lxml_1603216366346/work
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==4.0.3
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///C:/ci/nbconvert_1601914925608/work
nbformat @ file:///tmp/build/80754af9/nbformat_1610738111109/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
notebook @ file:///C:/ci/notebook_1611348396724/work
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.44
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandas==1.2.3
pandocfilters @ file:///C:/ci/pandocfilters_1605102497129/work
parso @ file:///tmp/build/80754af9/parso_1607623074025/work
path==15.1.0
path.py==12.5.0
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///C:/ci/pillow_1609786840597/work
pluggy==0.13.1
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1606344362066/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.13.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyfiglet==0.8.post1
Pygments @ file:///tmp/build/80754af9/pygments_1613849567472/work
pylint==2.6.0
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pypiwin32==223
pyrsistent @ file:///C:/ci/pyrsistent_1600141795814/work
pytest==6.2.2
pytest-shutil==1.7.0
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
python-resize-image==1.1.19
pytz==2020.1
pywin32==300
pywinpty==0.5.7
pyzmq==20.0.0
qtconsole==5.0.2
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scenedetect==0.5.4.1
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.1
selenium==3.141.0
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
sip==4.19.13
six @ file:///C:/ci/six_1605187374963/work
sklearn==0.0
soupsieve==2.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.2
testpath @ file:///home/ktietz/src/ci/testpath_1611930608132/work
threadpoolctl==2.1.0
toml==0.10.2
tornado @ file:///C:/ci/tornado_1606942392901/work
tqdm==4.51.0
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.11
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
windows-curses==2.2.0
wrapt==1.12.1
zipp==3.4.1
